+++ geo-types/src/geometry/coord.rs
@@ -319,10 +319,50 @@ where
     }
 }
 
-#[cfg(feature = "rstar")]
-impl<T> ::rstar::Point for Coord<T>
+// TODO: rstar_0_8 is deprecated.
+// this test should be deleted  before the release of version 0.8.0
+#[cfg(feature = "rstar_0_8")]
+impl<T> ::rstar_0_8::Point for Coord<T>
 where
-    T: ::num_traits::Float + ::rstar::RTreeNum,
+    T: ::num_traits::Float + ::rstar_0_8::RTreeNum,
+{
+    type Scalar = T;
+
+    const DIMENSIONS: usize = 2;
+
+    #[inline]
+    fn generate(generator: impl Fn(usize) -> Self::Scalar) -> Self {
+        coord! {
+            x: generator(0),
+            y: generator(1),
+        }
+    }
+
+    #[inline]
+    fn nth(&self, index: usize) -> Self::Scalar {
+        match index {
+            0 => self.x,
+            1 => self.y,
+            _ => unreachable!(),
+        }
+    }
+
+    #[inline]
+    fn nth_mut(&mut self, index: usize) -> &mut Self::Scalar {
+        match index {
+            0 => &mut self.x,
+            1 => &mut self.y,
+            _ => unreachable!(),
+        }
+    }
+}
+
+// TODO: rstar_0_9 is deprecated.
+// this test be adjusted before the release of version 0.8.0
+#[cfg(feature = "rstar_0_9")]
+impl<T> ::rstar_0_9::Point for Coord<T>
+where
+    T: ::num_traits::Float + ::rstar_0_9::RTreeNum,
 {
     type Scalar = T;