+++ geo-types/src/geometry/line.rs
@@ -221,7 +221,9 @@ impl<T: AbsDiffEq<Epsilon = T> + CoordNum> AbsDiffEq for Line<T> {
     }
 }
 
-#[cfg(feature = "rstar")]
+// TODO: rstar_0_9 and rstar_0_8 are deprecated.
+// This line should be adjusted  before the release of version 0.8.0
+#[cfg(any(feature = "rstar_0_8", feature = "rstar_0_9"))]
 macro_rules! impl_rstar_line {
     ($rstar:ident) => {
         impl<T> ::$rstar::RTreeObject for Line<T>
@@ -248,8 +250,15 @@ macro_rules! impl_rstar_line {
     };
 }
 
-#[cfg(feature = "rstar")]
-impl_rstar_line!(rstar);
+// TODO: rstar_0_8 is deprecated.
+// This line should be removed before the release of version 0.8.0
+#[cfg(feature = "rstar_0_8")]
+impl_rstar_line!(rstar_0_8);
+
+// TODO: rstar_0_9 is deprecated.
+// This line should be updated before the release of version 0.8.0
+#[cfg(feature = "rstar_0_9")]
+impl_rstar_line!(rstar_0_9);
 
 #[cfg(test)]
 mod test {