+++ geo-types/src/geometry/point.rs
@@ -586,10 +586,44 @@ where
     }
 }
 
-#[cfg(feature = "rstar")]
-impl<T> ::rstar::Point for Point<T>
+// TODO: rstar_0_8 is deprecated.
+// This should be removed  before the release of version 0.8.0
+#[cfg(feature = "rstar_0_8")]
+// These are required for rstar RTree
+impl<T> ::rstar_0_8::Point for Point<T>
 where
-    T: ::num_traits::Float + ::rstar::RTreeNum,
+    T: ::num_traits::Float + ::rstar_0_8::RTreeNum,
+{
+    type Scalar = T;
+
+    const DIMENSIONS: usize = 2;
+
+    fn generate(generator: impl Fn(usize) -> Self::Scalar) -> Self {
+        Point::new(generator(0), generator(1))
+    }
+
+    fn nth(&self, index: usize) -> Self::Scalar {
+        match index {
+            0 => self.0.x,
+            1 => self.0.y,
+            _ => unreachable!(),
+        }
+    }
+    fn nth_mut(&mut self, index: usize) -> &mut Self::Scalar {
+        match index {
+            0 => &mut self.0.x,
+            1 => &mut self.0.y,
+            _ => unreachable!(),
+        }
+    }
+}
+
+// TODO: rstar_0_9 is deprecated.
+// This line should be updated before the release of version 0.8.0
+#[cfg(feature = "rstar_0_9")]
+impl<T> ::rstar_0_9::Point for Point<T>
+where
+    T: ::num_traits::Float + ::rstar_0_9::RTreeNum,
 {
     type Scalar = T;