+++ geo-types/src/lib.rs
@@ -63,7 +63,12 @@
 //! - `approx`: Allows geometry types to be checked for approximate equality with [approx]
 //! - `arbitrary`: Allows geometry types to be created from unstructured input with [arbitrary]
 //! - `serde`: Allows geometry types to be serialized and deserialized with [Serde]
-//! - `use-rstar`: Allows geometry types to be inserted into [rstar] R*-trees
+//! - 'rstar' Allows geometry types to be inserted into [rstar] R*-trees.
+//! - `use-rstar_0_8`: (deprected) Allows geometry types to be inserted into [rstar] R*-trees (`rstar v0.8`)
+//! - `use-rstar_0_9`: (dprected) Allows geometry types to be inserted into [rstar] R*-trees (`rstar v0.9`)
+//!
+//!  The features "use-rstar_0_8" and "use-rstar_0_9" are deprected in 0.7.9 and will be removed in 0.8.0
+//!  Migration Guide: use feature "rstar" as a directl replacement.
 //!
 //! [approx]: https://github.com/brendanzab/approx
 //! [arbitrary]: https://github.com/rust-fuzz/arbitrary
@@ -81,6 +86,10 @@ use std::fmt::Debug;
 #[cfg(feature = "serde")]
 #[macro_use]
 extern crate serde;
+
+// TODO: rstar_0_8 is deprected, removed this before the release of version 0.8.0.
+#[cfg(feature = "rstar_0_8")]
+extern crate rstar_0_8;
 
 #[cfg(test)]
 #[macro_use]
@@ -123,7 +132,7 @@ mod macros;
 #[cfg(feature = "arbitrary")]
 mod arbitrary;
 
-#[cfg(feature = "rstar")]
+#[cfg(any(feature = "rstar_0_8", feature = "rstar_0_9"))]
 #[doc(hidden)]
 pub mod private_utils;
 
@@ -203,12 +212,29 @@ mod tests {
         assert_eq!(p.x(), 1_000_000i64);
     }
 
-    #[cfg(feature = "rstar")]
+    // Deprecation rstar_0_8 is deprecated.
+    // This test should be removed before the release of version 0.8.0.
+    #[cfg(feature = "rstar_0_8")]
     #[test]
     /// ensure Line's SpatialObject impl is correct
-    fn line_test_0_9() {
-        use rstar::primitives::Line as RStarLine;
-        use rstar::{PointDistance, RTreeObject};
+    fn line_test() {
+        use rstar_0_8::primitives::Line as RStarLine;
+        use rstar_0_8::{PointDistance, RTreeObject};
+
+        let rl = RStarLine::new(Point::new(0.0, 0.0), Point::new(5.0, 5.0));
+        let l = Line::new(coord! { x: 0.0, y: 0.0 }, coord! { x: 5., y: 5. });
+        assert_eq!(rl.envelope(), l.envelope());
+        // difference in 15th decimal place
+        assert_relative_eq!(26.0, rl.distance_2(&Point::new(4.0, 10.0)));
+        assert_relative_eq!(25.999999999999996, l.distance_2(&Point::new(4.0, 10.0)));
+    }
+
+    #[cfg(any(feature = "rstar_0_9", feature = "rstar"))]
+    #[test]
+    /// ensure Line's SpatialObject impl is correct
+    fn line_test() {
+        use rstar_0_9::primitives::Line as RStarLine;
+        use rstar_0_9::{PointDistance, RTreeObject};
 
         let rl = RStarLine::new(Point::new(0.0, 0.0), Point::new(5.0, 5.0));
         let l = Line::new(coord! { x: 0.0, y: 0.0 }, coord! { x: 5., y: 5. });